refactor: share skill command formatters
